Add get_circuit_assignments
This function converts named circuit assignments to assignments of
vamp-ir variableIds.

This will be used in the library proving API when the user supplies the
variable assignemnts associated with a circuit.

This new function returns an Error if a variable assignment is missing
for the circuit or the value of an assignment is invalid. Previously the
code would panic in these circumstances.