verify does not require the circuit
The proof API now returns:

```
pub struct ProofDataHalo2 {
    pub proof: Vec<u8>,
    pub verifying_key: VerifyingKey<EqAffine>,
    pub public_fields: Vec<String>,
    pub params: Params<EqAffine>,
}
```

The proof, verifying_key and params are used in the Halo2 verification
API. The public fields are a vector of vamp-ir public variable names
that need to be provided for verification. The order of these Strings is
used to construct the correct ordered vector of field elements to pass
to the verifier.

The verify API is now:

```
pub fn verify(
    proof_data: &ProofDataHalo2,
    named_public_assignments: &HashMap<impl AsRef<str>, Fp>,
    _config: &Config,
) -> Result<(), Error>
```

The circuit is not required for verification.